requirements.txt
Flask==2.1.1
python-dotenv==0.19.2
pymongo==4.2.0
Werkzeug==2.1.0
pymongo[srv]==4.2.0

